refactor: UIImageView 파일로 분리
