+++ OpenMarket/OpenMarket/ViewController/GridVC.swift
@@ -56,17 +56,5 @@ extension GridVC: UICollectionViewDataSource {
     }
 }
 
-extension String {
-    func strikeThrough() -> NSAttributedString {
-        let attributeString = NSMutableAttributedString(string: self)
-        attributeString.addAttribute(NSAttributedString.Key.strikethroughStyle, value: NSUnderlineStyle.single.rawValue, range: NSMakeRange(0,attributeString.length))
-        return attributeString
-    }
-    
-    func removeStrikeThrough() -> NSAttributedString {
-        let attributeString = NSMutableAttributedString(string: self)
-        attributeString.removeAttribute(NSAttributedString.Key.strikethroughStyle, range: NSMakeRange(0, attributeString.length))
-        return attributeString
-    }
-}
+
